Add FeedPerson model
The FeedPerson model includes a str method that is the full name of the FeedPerson instance. The test that was added in tests.py checks that this method is working correctly. The max length and nullability of each field is not being tested, because this should be guaranteed by the framework.