Add Person model
The Person model has the fields firstname, lastname, location, and name. The str method returns the value of the name field. Each of these fields are not nullable. A meta class was added to the model, because without specifying the verbose_name_plural, the model in Django Admin was showing up as "Persons." By setting verbose_name_plural to "People," Django Admin lists "People" instead.