Add Rollup function and tests
The Rollup function deletes all of the current rollup objects and creates new rollup objects by looping over the Place, Person, and FeedPerson objects and using their name and location field values. The tests check that the correct data is in the final rollup objects and that the original data was deleted.