Add load_feed_people function
This function removes all of the existing FeedPerson data before importing updated data from the active SEAS feed. This uses BeautifulSoup to parse the data, filters for people with a location of 114 Western Ave or 150 Western Ave, and generates new FeedPerson data objects with the keys eppn, firstname, lastname, name, and location. Then it calls the load_rollup function to generate Rollup objects from the new FeedPerson objects and returns an HttpResponse to indicate that the data import was successful.