Prevent indexing of data when running tests
The data was originally being indexed when tests were run. Algolia seemed to be differentiating records based off of primary keys, and duplicate records were being added in Algolia. As a fix, we are using disable_auto_indexing to prevent indexing when running the tests: https://github.com/algolia/algoliasearch-django#temporarily-disable-the-auto-indexing