Rename dev compose file
As a minor quality of life improvement, this renames the
docker-compose.dev.yml to just docker-compose.yml so you no longer need
to pass the --file arugment to the cli. It also changes the build stage
to use our new unified Dockerfile with the appropriate target.

I've left the docker-compose.prod.yml here, updating the
"dockerfile:" and "command:" fields as necesary. As we'll be running the
app on ECS we won't need this for "actual" production, but it could be
useful to simulate a production environment in the future.