Set ownership of media/static dirs
When I tried bringing up a fresh build of the production container I
started getting errors about missing permissions on the media and static
directories, which makes sense as the volumes are being mounted in the
nginx container as root-owned but in the django container as app-owned.
This adds an app user to the nginx container and gives that user owner
of media and static, so the permissions will match across both
containers.