Use a single .env file
This is not a major change, but the .env file will be read automatically
by docker-compose without needing to be specified with the `env_file`
option, making it just that little bit easier to get started quickly.

I've moved the hard-coded differences between the two files (i.e DEBUG
and DJANGO_SETTINGS_MODULE`) directly into the docker-compose file so
that we don't need to maintain the separate templates, and removed the
POSTGRES_ specific environment variables in favor or defining them from
the SQL_ variables in the docker-compose. In development we need to have
the same user, password, and database defined in both, and this just
helps avoid any typos that could cause things to fail.