Add Docker Documentation to the README
This provides a brief overview of the application as a whole and it's
connection to Algolia and to the sec-directory-client repo, as well as
more detailed instructions on how to get up and running initially.

Missing are the details about running the load_feed_people function and
indexing the data.