Add tests to check rendering and result of clicking the "Update Screens" button in Django Admin
The "Update Screens" button exists under People and Places in Django Admin. These tests make sure that the button renders under each. Also, it checks that when we follow the link that updates the screens (in place of simulating a click event) that we are able to see the success message in Django Admin.