Add exec option
This changes from a boolean PRODUCTION value to an EXEC_MODE value that
can be set to production or development, but defaults to "exec". In exec
mode, it will run whatever is passed in as the command value so that we
can run the load_feed_people() function or other django-admin commands.
In the other modes, it will run the app in either production or
development mode as indicated, after running migrations and creating the
superuser.