Condense PlaceAdmin and PersonAdmin into UpdateScreens
Instead of having two classes that are essentially the same, condense this into one class, UpdateScreens, which lives in the rollup utils file. This also adds a try catch clause in case something goes wrong with load_rollup. If there is an exception, it is logged and an error message is shown to the user in Django Admin.