Update base container to python:3.8-alpine
The python:3.8.0-alpine container is using an EOL version of alpine
(3.10.3), but dropping the .0 patch level will make it use the most
up-to-date versions of both python (3.8.15) and alpine (3.16.3). Those
versions go EOL in October 2024 and May 2024, respectively, so that does
give us some runway to set a more formal and consistent upgrade process.